+++ resources/lang/ca/documents.php
@@ -21,5 +21,6 @@ return [
     'empty' => 'No hi ha documents en aquesta secció',
     'is_task' => 'És una tasca?',
     'mark_done' => 'Marca com a fet. ',
+    'done' => 'Realitzat',
 
 ];