+++ resources/lang/ca/students.php
@@ -90,5 +90,6 @@ return [
     'log_card_assign' => 'Carta assignada',
     'log_pet' => 'Mascota',
     'log_skill' => 'Resultat d\'una habilitat teva o d\'algú.',
+    'log_task' => 'Has fet una tasca',
 
 ];