+++ resources/lang/en/documents.php
@@ -21,5 +21,6 @@ return [
     'empty' => 'There are no documents in this section',
     'is_task' => 'Is a task?',
     'mark_done' => 'Mark as done. ',
+    'done' => 'Done',
 
 ];