+++ resources/lang/es/documents.php
@@ -21,5 +21,6 @@ return [
     'empty' => 'No hay documentos en esta sección',
     'is_task' => '¿Es una tarea?',
     'mark_done' => 'Marca como hecho. ',
+    'done' => 'Realizado',
 
 ];