+++ resources/lang/es/students.php
@@ -90,5 +90,6 @@ return [
     'log_card_assign' => 'Carta asignada',
     'log_pet' => 'Mascota',
     'log_skill' => 'Resultado de una habilidad, tuya o de alguien.',
+    'log_task' => 'Has realizado una tarea',
 
 ];